requirements.txt
certifi==2024.8.30
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
cryptography==43.0.1
filelock==3.12.2
Flask==2.2.5
fsspec==2023.1.0
future==1.0.0
gevent==22.10.2
git-filter-repo==2.45.0
greenlet==3.1.1
huggingface-hub==0.16.4
idna==3.10
importlib-metadata==6.7.0
itsdangerous==2.1.2
Jinja2==3.1.4
MarkupSafe==2.1.5
numpy==1.21.6
packaging==24.0
pdfminer.six==20231228
pycparser==2.21
PySocks==1.7.1
PyYAML==6.0.1
regex==2024.4.16
requests==2.31.0
requests-file==2.1.0
safetensors==0.4.5
tldextract==4.0.0
tokenizers==0.13.3
torch==1.13.1
tqdm==4.66.5
transformers==4.30.2
typing_extensions==4.7.1
urllib3==2.0.7
Werkzeug==2.2.3
zipp==3.15.0
zope.event==5.0
zope.interface==6.4.post2
